Moved Responder tests to a separate class. Resolves #16.
Moved the tests of the git responder and the e-mail responder to a separate class, ResponderTest.java. Also uncommented the error output of send() in Responder.java.
